requirements.txt
beautifulsoup4~=4.12.2
Scrapy~=2.9.0
crawler~=0.0.2
requests~=2.31.0
PyYAML~=6.0